// Portal.Portal.Web/src/app/shared/infrastructure/PortalHttpClient/model/resetPasswordRequest.ts
/**
 * Demo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ResetPasswordRequest { 
    password?: string;
    confirmPassword?: string;
    email?: string;
    token?: string;
}